requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
filelock==3.16.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
pillow==10.2.0
pycparser==2.22
regex==2024.9.11
requests==2.32.3
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
sounddevice==0.5.1
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
